refactor from app to native_image
