+++ shell/browser/api/electron_api_app.cc
@@ -47,7 +47,6 @@
 #include "shell/common/electron_paths.h"
 #include "shell/common/gin_converters/callback_converter.h"
 #include "shell/common/gin_converters/file_path_converter.h"
-#include "shell/common/gin_converters/gfx_converter.h"
 #include "shell/common/gin_converters/gurl_converter.h"
 #include "shell/common/gin_converters/image_converter.h"
 #include "shell/common/gin_converters/net_converter.h"
@@ -1569,9 +1568,6 @@ gin::ObjectTemplateBuilder App::GetObjectTemplateBuilder(v8::Isolate* isolate) {
       .SetMethod(
           "getApplicationInfoForProtocol",
           base::BindRepeating(&Browser::GetApplicationInfoForProtocol, browser))
-      .SetMethod(
-          "createThumbnailFromPath",
-          base::BindRepeating(&Browser::CreateThumbnailFromPath, browser))
 #endif
       .SetMethod(
           "getApplicationNameForProtocol",
